Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
 
+import services.exceptions.DatabaseEoxception;
 import services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -27,9 +30,19 @@ public class UserService {
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
+
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+	    try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseEoxception(e.getMessage());		
+	    }	
+	} 
+	
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
